Registration form

+++ base/views.py
@@ -4,8 +4,9 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
-# from django.http import HttpResponse
-# We get rid of HttResponse after adding our template in Settings and use render
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 from .forms import RoomForm
 # Create your views here.
 # rooms = [
@@ -15,6 +16,10 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
         user_name = request.POST.get('username')
         user_password = request.POST.get('password')
@@ -30,12 +35,20 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request, 'Username or password does not exist.')
-    context= {}
+    context= {'page':page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    
+    form = UserCreationForm()
+    context = {'form': form}
+    return render(request, 'base/login_register.html', context)
+
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
@@ -54,6 +67,8 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+# Restricting unauthorized user from accessing some pages.
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -66,9 +81,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    # user only update the room if he's the host.
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -78,8 +97,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def delete(request, pk):
     room = Room.objects.get(id=pk)
+     # user only delete the room if he's the host.
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+    
     if request.method == "POST":
         room.delete()
         return redirect('home')
